+++ examples/voxels/assets/voxel.vert
@@ -1,6 +1,6 @@
 #version 410 core
 
-layout (location = 0) in ivec3 aPos;
+layout (location = 0) in vec3 aPos;
 layout (location = 1) in vec2 aTex;
 layout (location = 2) in vec4 aColor;
